lab8/gym/requirements.txt
tensorforce[tf_gpu]
pyyaml
grpcio-tools
gym
numpy
